🎨 하단 리스트 수정
